Bunch o' fixes
