+++ lib/config.js
@@ -16,6 +16,7 @@ module.exports = {
     DS_CRYPT_KEY_SEED: 'MercedesRedBullRacingHondaMcLarenRenaultRacingPointBWTMercedesRenaultFerrariAlphaTauriHondaAlfaRomeoRacingFerrariHaasFerrariWilliamsMercedes',
     DS_CRYPT_ALGO: 'aes-192-cbc',
     HEADLESS: true,
+    ENTITLEMENT: 'F1_TV_Pro_Annual',
 
     makeItGreen: (str) => {
         return '\x1b[32m' + str + '\x1b[37m';
