+++ lib/f1tv-api.js
@@ -2,90 +2,60 @@ const config = require('./config');
 const axios = require('axios');
 const DataStore = require('./secure-local-data');
 const f1tvToken = require('./f1tv-token');
+const { isF1tvUrl } = require('./f1tv-validator');
 
-const getSlugName = (urlStr) => {
-    try { return new URL(urlStr).pathname.split('/').pop(); }
-    catch (e) { return undefined; }
-};
+const getAdditionalStreamInfo = (streams, searchStr) => {
+    for ( const stream of streams ) {
+        //const data = (stream.type === 'obc')?`name: ${config.makeItGreen(stream.driverFirstName+' '+stream.driverLastName)}`.padEnd(37) + `number: ${config.makeItGreen(stream.racingNumber)}`.padEnd(22) + `tla: ${config.makeItGreen(stream.title)}`:`name: ${config.makeItGreen(stream.title)}`;
 
-const getEpisodeUrl = (urlStr) => {
-    const slug = getSlugName(urlStr);
-    const options = {
-        baseURL: config.BASE_URL,
-        params: {
-            'fields': 'items,items__ovps,items__ovps__vod_id,items__self',
-            'slug': slug
-        }
-    };
-    return axios.get('/api/episodes/', options)
-        .then(response => {
-            return response.data.objects.shift().items.shift();
-        })
-        .catch(e => e);
-};
-
-const getSessionUrl = (urlStr, searchStr = 'wif') => {
-    return getSessionOccuurrence(urlStr)
-        .then(id => getSessionChannels(id))
-        .then(channels => getSessionChannelUrl(searchStr, channels));
-};
-
-const getSessionOccuurrence = (urlStr) => {
-    const slug = getSlugName(urlStr);
-    const options = {
-        baseURL: config.BASE_URL,
-        params: {
-            'fields': 'availability_details,status,uid',
-            'slug': slug
-        }
+        // Need to create a search that finds the right additional stream based on searchStr
     }
-    return axios.get('/api/session-occurrence/', options)
-        .then(response => response.data.objects.shift().uid)
 };
 
-const getSessionChannels = (id) => {
-    const options = {
-        baseURL: config.BASE_URL,
-        params: {
-            'fields': 'channel_urls'
-        }
-    }
-    return axios.get(`/api/session-occurrence/${id}/`, options)
-        .then(response => response.data.channel_urls);
+const getChannelIdFromPlaybackUrl = (url) => {
+    const path = new URL(config.BASE_URL + url);
+    return path.searchParams.get('channelId');
 };
 
-const getSessionChannelUrl = (searchStr, channels = []) => {
-    const channel = channels.shift();
-    const options = {
-        baseURL: config.BASE_URL,
-        params: {
-            'fields_to_expand': 'driveroccurrence_urls'
-        }
+const getContentParams = (url) => {
+    if (!isF1tvUrl(url)) throw new Error("Invalid F1TV URL");
+    const content = new URL(url);
+    const temp = content.pathname.split('/');
+    return { 
+        name: temp.pop(),
+        id: temp.pop()
     };
+};
 
-    return axios.get(channel, options)
-        .then(response => {
-            if (response.data.name === undefined) { throw new Error('No channel matches ' + config.makeItGreen(searchStr)); }
-            const data = (response.data.channel_type === 'driver') ? [response.data.name, response.data.driveroccurrence_urls[0].driver_tla, `${response.data.driveroccurrence_urls[0].driver_racingnumber}`] : [response.data.name];
-            if (data.find(item => item.toLowerCase().indexOf(searchStr.toLowerCase()) !== -1) !== undefined) { return response.data.self }
-            return getSessionChannelUrl(searchStr, channels);
-        })
-}
+const getContentInfo = (url) => {
+    const content = getContentParams(url);
+    const options = {
+        baseURL: config.BASE_URL,
+    }
+    const path = '/2.0/R/ENG/BIG_SCREEN_HLS/ALL/CONTENT/VIDEO/' + content.id + '/' + config.ENTITLEMENT + '/2';
 
-const getFinalUrl = async (itemPath) => {
+    return axios.get(path, options)
+        .then(result => result.data.resultObj.containers.shift());
+};
+
+const getContentStreamUrl = (id , channel=null) => {
     const ds = new DataStore(`${config.HOME}${config.PATH_SEP}${config.DS_FILENAME}`);
-    const token = ds.get('token')
-    return getTokenizedUrl(itemPath, token);
- };
+    const token = ds.get('token');
+    const params = (channel === null)?{ 'contentId': id }:{ 'contentId': id, 'channelId': channel };
+    const options = {
+        baseURL: config.BASE_URL,
+        headers: {
+            'entitlementtoken': token
+        },
+        params: params
+    }
+    const path = '/1.0/R/ENG/BIG_SCREEN_HLS/ALL/CONTENT/PLAY';
 
-const getTokenizedUrl = (itemPath, jwt) => {
-    const isAsset = (itemPath.indexOf('assets') !== -1);
-    const item = (isAsset) ? { 'asset_url': itemPath } : { 'channel_url': itemPath };
-    const authHeader = { 'Authorization': `JWT ${jwt}` };
+    console.log(options);
 
-    return axios.post('/api/viewings/', item, { baseURL: config.BASE_URL, headers: authHeader })
-        .then(response => (isAsset) ? response.data.objects.shift().tata.tokenised_url : response.data.tokenised_url);
-}
+    return axios.get(path, options)
+        .then(result => result.data.resultObj.url);
+};
 
 const saveF1tvToken = async (user, pass) => {
     const ds = new DataStore(`${config.HOME}${config.PATH_SEP}${config.DS_FILENAME}`);
@@ -96,12 +66,9 @@ const saveF1tvToken = async (user, pass) => {
 
 
 module.exports = {
-    getSlugName,
-    getEpisodeUrl,
-    getSessionOccuurrence,
-    getSessionChannels,
-    getSessionUrl,
-    getFinalUrl,
-    getTokenizedUrl,
+    getContentParams,
+    getContentInfo,
+    getContentStreamUrl,
+    getChannelIdFromPlaybackUrl,
     saveF1tvToken    
 };
