+++ lib/f1tv-validator.js
@@ -11,17 +11,7 @@ const isF1tvUrl= (urlStr) => {
         if (isUrl(urlStr)) {
             let url = new URL(urlStr);
             return url.host.toLowerCase().indexOf('f1tv') !== -1 &&
-                ['current-season', 'archive', 'episode'].find(item => url.pathname.toLowerCase().indexOf(item) !== -1);
-        }
-        return false;
-    }
-    catch (e) { return false; }
-}
-
-const isF1tvEpisodeUrl = (urlStr) => {
-    try {
-        if (isUrl(urlStr)) {
-            return new URL(urlStr).pathname.toLowerCase().indexOf('episode') !== -1;
+                ['detail'].find(item => url.pathname.toLowerCase().indexOf(item) !== -1);
         }
         return false;
     }
@@ -30,6 +20,5 @@ const isF1tvEpisodeUrl = (urlStr) => {
 
 module.exports = {
     isUrl,
-    isF1tvUrl,
-    isF1tvEpisodeUrl
+    isF1tvUrl
 };
